made updates in unit tests accordingly
