+++ Services/MainApp/MainApp.Test/Application/Service.Tests/ProblemsService.Tests.cs
@@ -1,5 +1,7 @@
-﻿using BuildingBlocks.Common.Enums;
+﻿using BuildingBlocks.Common.Dtos;
+using BuildingBlocks.Common.Enums;
 using FluentAssertions;
+using MainApp.Application.Dto.Request;
 using MainApp.Application.Services;
 using MainApp.Domain.Entity;
 using MainApp.Infrastructure.Data;
@@ -163,6 +165,135 @@ public class ProblemsServiceTests
         result.Should().HaveCount(2); 
         result[0].AuthUsername.Should().Be("User1"); 
         result[1].AuthUsername.Should().Be("User2");
+    }
+    
+      [Fact]
+    public async Task AddProblem_ShouldAddProblem_WhenProblemIsValid()
+    {
+        // Arrange
+        var problemDto = new AddProblemDto
+        {
+            Name = "New Problem",
+            ProblemText = "Test Problem",
+            Difficulty = "Easy",
+            Categories = ["Arrays","Strings"],
+            RuntimeLimit = 2000,
+            MemoryLimit = 256,
+            TestCases = new List<TestCaseDto>
+            {
+                new TestCaseDto { Input = "1 2", ExpectedOutput = "3" }
+            }
+        };
+
+        // Act
+        await _problemsService.AddProblem(problemDto);
+
+        // Assert
+        var problemInDb = await _fakeDbContext.Problems.FirstOrDefaultAsync(p => p.Name == "New Problem");
+        problemInDb.Should().NotBeNull();
+        problemInDb.Name.Should().Be("New Problem");
+        problemInDb.TestCases.Should().HaveCount(1);
+    }
+
+    [Fact]
+    public async Task EditProblem_ShouldEditProblem_WhenProblemExists()
+    {
+        // Arrange
+        var problem = new Problem
+        {
+            Id = 1,
+            Name = "Old Problem",
+            ProblemText = "Old text",
+            Difficulty = Difficulty.Easy,
+            Category = Category.Arrays,
+            RuntimeLimit = 500,
+            MemoryLimit = 128,
+            TestCases = new List<TestCase>()
+        };
+
+        _fakeDbContext.Problems.Add(problem);
+        await _fakeDbContext.SaveChangesAsync();
+
+        var editProblemDto = new AddProblemDto
+        {
+            Name = "Updated Problem",
+            ProblemText = "Problem Text",
+            Difficulty = "Easy",
+            Categories = ["Arrays","Strings"],
+            RuntimeLimit = 1000,
+            MemoryLimit = 256,
+            TestCases = new List<TestCaseDto>
+            {
+                new TestCaseDto { Input = "1 3", ExpectedOutput = "4" }
+            }
+        };
+
+        // Act
+        await _problemsService.EditProblem(1, editProblemDto);
+
+        // Assert
+        var updatedProblem = await _fakeDbContext.Problems.FirstOrDefaultAsync(p => p.Id == 1);
+        updatedProblem.Should().NotBeNull();
+        updatedProblem.Name.Should().Be("Updated Problem");
+        updatedProblem.Difficulty.Should().Be(Difficulty.Easy);
+        updatedProblem.TestCases.Should().HaveCount(1);
+    }
+
+    [Fact]
+    public async Task EditProblem_ShouldThrowException_WhenProblemDoesNotExist()
+    {
+        // Arrange
+        var editProblemDto = new AddProblemDto
+        {
+            Name = "Updated Problem",
+            ProblemText = "Problem Text",
+            Difficulty = "Easy",
+            Categories = ["Arrays","Strings"],
+            RuntimeLimit = 1000,
+            MemoryLimit = 256,
+            TestCases = new List<TestCaseDto>()
+        };
+
+        // Act & Assert
+        var act = async () => await _problemsService.EditProblem(999, editProblemDto);
+        await act.Should().ThrowAsync<InvalidOperationException>()
+            .WithMessage("Problem not found");
+    }
+
+    [Fact]
+    public async Task DeleteProblem_ShouldRemove_WhenProblemExists()
+    {
+        // Arrange
+        var problem = new Problem
+        {
+            Id = 1,
+            Name = "Test Problem",
+            ProblemText = "Test problem text",
+            Difficulty = Difficulty.Medium,
+            Category = Category.Arrays,
+            RuntimeLimit = 1000,
+            MemoryLimit = 256,
+            TestCases = new List<TestCase>()
+        };
+
+        _fakeDbContext.Problems.Add(problem);
+        await _fakeDbContext.SaveChangesAsync();
+
+        // Act
+        await _problemsService.DeleteProblem(1);
+
+        // Assert
+        var deletedProblem = await _fakeDbContext.Problems.FirstOrDefaultAsync(p => p.Id == 1);
+        deletedProblem.Should().BeNull();
+    }
+
+    [Fact]
+    public async Task DeleteProblem_ShouldThrowException_WhenProblemDoesNotExist()
+    {
+        // Act & Assert
+        var act = async () => await _problemsService.DeleteProblem(999);
+        await act.Should().ThrowAsync<InvalidOperationException>()
+            .WithMessage("Problem not found");
     }
 
     public async Task GetSubmissions_ShouldThrowException_WhenProblemDontExists()
